Extract visibility change triggers
